move tests out of autoloader
