+++ composer.json
@@ -39,9 +39,22 @@
         "psr/log": "^1.1",
         "symfony/cache": "^4.0"
     },
+    "require-dev": {
+        "clue/graph-composer": "dev-master",
+        "mf2/tests": "@dev",
+        "mindplay/composer-locator": "^2.1",
+        "phpunit/phpunit": ">=7.0",
+        "php-coveralls/php-coveralls": "^2.1",
+        "squizlabs/php_codesniffer": "^3.3"
+    },
     "autoload": {
         "psr-4": {
             "Jkphl\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Jkphl\\": "tests/"
         }
     },
     "scripts": {
@@ -50,13 +63,5 @@
         "check-style": "vendor/bin/phpcs -p --standard=PSR2 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src",
         "fix-style": "vendor/bin/phpcbf -p --standard=PSR2 --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1 src",
         "test": "@phpunit"
-    },
-    "require-dev": {
-        "clue/graph-composer": "dev-master",
-        "mf2/tests": "@dev",
-        "mindplay/composer-locator": "^2.1",
-        "phpunit/phpunit": ">=7.0",
-        "php-coveralls/php-coveralls": "^2.1",
-        "squizlabs/php_codesniffer": "^3.3"
     }
 }
