+++ tests/Micrometa/Fixture/microformats/entry.html
