+++ tests/Micrometa/Fixture/mixed/valid-with-errors-test.html
