+++ tests/Micrometa/Fixture/rdfa-lite/article-rdfa-lite.html
