+++ tests/Micrometa/Ports/ItemObjectModelTest.php
@@ -34,7 +34,7 @@
  *  CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  ***********************************************************************************/
 
-namespace Jkphl\Micrometa\Tests\Ports;
+namespace Jkphl\Micrometa\Ports;
 
 use Jkphl\Micrometa\Application\Factory\PropertyListFactory;
 use Jkphl\Micrometa\Application\Item\Item as ApplicationItem;
@@ -46,8 +46,8 @@ use Jkphl\Micrometa\Ports\Item\Item;
 use Jkphl\Micrometa\Ports\Item\ItemInterface;
 use Jkphl\Micrometa\Ports\Item\ItemList;
 use Jkphl\Micrometa\Ports\Item\ItemObjectModel;
-use Jkphl\Micrometa\Tests\AbstractTestBase;
-use Jkphl\Micrometa\Tests\MicroformatsFeedTrait;
+use Jkphl\Micrometa\AbstractTestBase;
+use Jkphl\Micrometa\MicroformatsFeedTrait;
 
 /**
  * Item object model tests
